Fix start signal timing
The previous version has a problem where the start signal is pulsed high
at the end of a read sequence. This causes unintuitive behaviour when
attempting to change the sensor's sensitivity using the method suggested in the
datasheet, which is to adjust the width of the initial start pulse. A
variable 'integrationTime' has been added to indicate how the sensitivity
should be changed, and the incorrectly timed start pulse has been
removed. The delay between the start signal going low and the start of
the reading of data has also been consolidated into one loop, and the
comment changed to better identify the delay's purpose.